Decouple query playground from HackerNews demo (#35)
This PR decouples the HackerNews-specific code from the query playground, which will allow us to create new demos against different data sources while reusing the same underlying components. Specifically, this PR introduces the TrustfallPlayground component which is data-agnostic, and the playgrounds/HackerNews component which uses that component to render the HN demo.