Support storing query and vars state in URL (#40)
This PR uses the fantastic use-query-params library to persist the query and vars strings in the URL, which makes our demos shareable! Since we are storing entire encoded queries in the URL, we can quickly run into URL length limits if we use the default functionality of storing data as query params. Fortunately, use-query-params allows us to specify a custom QueryParamAdapter that lets us convert to using hash fragments instead, which are not sent to the server!