Allow new Safari and iOS in the playground. (#524)
The two bugs blocking the HackerNews playground from working on Safari and iOS appear to have been fixed in a recent update. I've confirmed that the playground now works on Safari 16.6 on both desktop and iOS. Note that iOS 16.7 is the one that apparently ships with Safari 16.6, based on the device I was able to check with.