Add documentation based on Material for MkDocs. (#807)
This revives the add-docs branch, and will attempt to get it up to a
level for initial publishing.

Some things I've already done:

1. Move poetry to uv (I was going to skip this but I don't know how to
use poetry and there's a poetry to uv migrator so this was actually
fastest for me)
2. I filled in the last sections on the flutter schema guide

I need to figure out layout/linking/query-writing-guide. There should
probably be a implementers guide for doing the rust code. ~~Maybe I can
crib/steal some from my docker cleanup blogpost and adapt it to work
with a mock "flutter" server. I'll see when I get to it.~~

One thought, playground links for the trustfall playground that can be
embedded in the docs and people can click to go to the playground to run
them would be pretty good.

---------

Co-authored-by: Predrag Gruevski <obi1kenobi82@gmail.com>
Co-authored-by: Predrag Gruevski <2348618+obi1kenobi@users.noreply.github.com>